requirements.txt
openai 
python-pptx